Add KeyCallback, update window, modify variable scopes.
Add the KeyCallback and set it to get input from the window.
WindowShouldClose flag set to true when escape key is pressed. Update
the window and poll events to inform system that window is active.
Modify the scope to add multi-window support and make the methods more
independent on windowHandle global variable.