Add createShaders, compileShader, linkProgram, and shader source code.
Add createShaders to handle the full creation of shaders to prevent
redundant code.

Add the compile shader to compile the source code for the specified
shader, to be used by OpenGL.

Add the linkProgram to combine the shaders into one program for use by
OpenGL.

Add the shader source code which will tell OpenGL how the data such as
vertex coordinates and color is to be used.